requirements.txt
aiohttp
beautifulsoup4
lxml
requests
torch
transformers
pillow
numpy
supabase
python-dotenv